Fix bug for constructing Mesh<0>
Made the constructors for a Mesh<0> construct the same object as
such a Mesh is supposed to represent a single point (i.e. the
boundary Mesh of a Mesh<1>).

Previously, the consturctor taking isotropic_extents used its
value when setting the member variable quadrature_and_basis_,
while the other consturctors default initialized it.  This
caused the equivalence operator to fail.